Groupobjects with FK for Tag (#2272)

+++ ESSArch_Core/install/migrate_groupgeneric_to_groupfk.py
@@ -22,13 +22,15 @@
     Email - essarch@essolutions.se
 """
 import django
-from django.contrib.contenttypes.models import ContentType
 
 django.setup()
+
+from django.contrib.contenttypes.models import ContentType  # noqa isort:skip
 
 from ESSArch_Core.tags.models import (  # noqa isort:skip
     StructureUnit,
     TagVersion,
+    Tag
 )
 from ESSArch_Core.ip.models import InformationPackage  # noqa isort:skip
 from ESSArch_Core.auth.models import Group, GroupGenericObjects  # noqa isort:skip
@@ -40,34 +42,43 @@ def migrate_gg(klass, group):
     ctype = ContentType.objects.get_for_model(klass)
     num = 0
     total_num = GroupGenericObjects.objects.filter(group=group, content_type=ctype).count()
+    if total_num:
+        print('group_obj: {} - ctype: {}'.format(group, ctype.name))
     for gg_obj in GroupGenericObjects.objects.filter(group=group, content_type=ctype):
-        go_obj, created = group.add_object(gg_obj.content_object)
-        num += 1
-        if created:
-            print('{} ({}) - {} object {} - added to group {}'.format(num,
-                  total_num, ctype.name, gg_obj.content_object, group))
+        if gg_obj.content_object:
+            go_obj, created = group.add_object(gg_obj.content_object)
+            num += 1
+            if created:
+                print('{} ({}) - {} object {} - added to group {}'.format(num,
+                                                                          total_num, ctype.name,
+                                                                          gg_obj.content_object, group))
+            else:
+                print('{} ({}) - {} object: {} - already related to group {}'.format(
+                    num, total_num, ctype.name, gg_obj.content_object, group))
         else:
-            print('{} ({}) - {} object: {} - already related to group {}'.format(
-                num, total_num, ctype.name, gg_obj.content_object, group))
+            print('{} ({}) - {} warning gg_obj {} with gg_obj.object_id {} does not have any content_object to \
+add to group {}'.format(num, total_num, ctype.name, gg_obj, gg_obj.object_id, group))
 
 
 def remove_gg(klass, group):
     ctype = ContentType.objects.get_for_model(klass)
     gg_objs = GroupGenericObjects.objects.filter(group=group, content_type=ctype)
     total_num = gg_objs.count()
-    print('Start to remove {} objects with with type: {} and group: {}'.format(total_num, ctype.name, group))
-    gg_objs.delete()
-    print('Finished to remove {} objects with with type: {} and group: {}'.format(total_num, ctype.name, group))
+    if total_num:
+        print('Start to remove {} objects with with type: {} and group: {}'.format(total_num, ctype.name, group))
+        gg_objs.delete()
+        print('Finished to remove {} objects with with type: {} and group: {}'.format(total_num, ctype.name, group))
 
 
 def migrate():
     for group_obj in Group.objects.filter(group_type__codename='organization'):
         print('group_obj: {}'.format(group_obj))
-        migrate(InformationPackage, group_obj)
-        migrate(AccessAid, group_obj)
-        migrate(Agent, group_obj)
-        migrate(TagVersion, group_obj)
-        migrate(StructureUnit, group_obj)
+        migrate_gg(InformationPackage, group_obj)
+        migrate_gg(AccessAid, group_obj)
+        migrate_gg(Agent, group_obj)
+        migrate_gg(Tag, group_obj)
+        migrate_gg(TagVersion, group_obj)
+        migrate_gg(StructureUnit, group_obj)
 
 
 def remove():
@@ -76,6 +87,7 @@ def remove():
         remove_gg(InformationPackage, group_obj)
         remove_gg(AccessAid, group_obj)
         remove_gg(Agent, group_obj)
+        remove_gg(Tag, group_obj)
         remove_gg(TagVersion, group_obj)
         remove_gg(StructureUnit, group_obj)
